+++ Castle_Kid/Assets/_Scripts/Projectiles/Projectile Scriptable Object.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 96eda53723fcdaa438c16f6f610b9a23, type: 3}
+  m_Name: Projectile Scriptable Object
+  m_EditorClassIdentifier: 
+  SparkCollider2D: {fileID: 0}
+  ConeSearchTargetCollider2D: {fileID: 0}
+  CircleSearchSenderCollider2D: {fileID: 0}
