+++ Castle_Kid/Assets/_Scripts/Projectiles/Projectile Scriptable Object.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 07408459e8cac8e4c92bc396bbccd7cd
-PrefabImporter:
+guid: 8cd9d17314af3f44292626dfb85bb16d
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
