+++ Castle_Kid/Assets/_Scripts/Projectiles/ProjectileScriptableObject.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 96eda53723fcdaa438c16f6f610b9a23
